requirements.txt
streamlit
sqlmodel
google-generativeai
python-dotenv
pydantic